+++ .github/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Release 0.41.0-dev
 
 ### New features since last release
+
 * Implement new C++ kernels for efficient in-place multiplication of sparse matrices to state vectors, supporting both controlled and non-controlled gates, and add comprehensive tests for this new functionality.
-  [#1085](https://github.com/PennyLaneAI/pennylane-lightning/pull/1085)
+  [(#1085)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1085)
+
+* Integrate sparse kernels into the Lightning-Qubit dynamic dispatcher, enhance the Python layers, and expand the testing process for dense and sparse operations.
+  [(#1094)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1094)
 
 * Add support for Python 3.13.
   [(#1001)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1001)